requirements.txt
telethon>=1.36.0
python-dotenv>=1.0.1
litellm>=1.55.0
tzdata>=2024.1
